# api/openapi.yaml
openapi: 3.0.0
info:
  title: Vector Search Service API
  description: |
    A vector search service providing semantic search capabilities using PostgreSQL + pgvector.
    
    This service provides document ingestion, embedding generation, and similarity search operations
    designed for integration with Elyra pipelines and other AI workflows.
    
    ## Features
    - Document ingestion with automatic chunking
    - Semantic similarity search
    - Batch processing for pipeline integration
    - Collection management for document organization
    - Asynchronous job tracking
    
    ## Authentication
    This service uses JWT tokens for authentication. Include the token in the Authorization header:
    `Authorization: Bearer <token>`
  version: 1.0.0
  contact:
    name: Vector Search Service
    url: https://github.com/vector-search-service
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://vector-search-service.openshift.apps.example.com
    description: OpenShift deployment
paths:
...
  /api/v1/search/similarity:
    post:
      summary: Semantic similarity search
      description: |
        Perform semantic similarity search on document embeddings.
        
        This endpoint:
        1. Generates an embedding for the query text
        2. Searches for similar embeddings in the vector database
        3. Returns the most similar documents with scores
      tags:
        - search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimilaritySearchRequest'
            examples:
              simple_search:
                summary: Simple text search
                value:
                  query: "How to configure the database connection?"
                  collection_id: "default"
                  limit: 5
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilaritySearchResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/search/batch:
    post:
      summary: Batch similarity search
      description: |
        Perform batch similarity searches for multiple queries.
        
        This endpoint is optimized for pipeline operations that need to
        search for multiple terms efficiently.
      tags:
        - search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSearchRequest'
      responses:
        '200':
          description: Batch search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSearchResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/search/collections:
    get:
      summary: List collections
      description: List all available collections for searching
      tags:
        - search
      responses:
        '200':
          description: List of collections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/jobs/{job_id}/status:
    get:
      summary: Get job status
      description: Get the status of an asynchronous job
      tags:
        - jobs
      parameters:
        - name: job_id
          in: path
          required: true
          description: Unique identifier of the job
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/jobs/{job_id}/results:
    get:
      summary: Get job results
      description: Retrieve the results of a completed job
      tags:
        - jobs
      parameters:
        - name: job_id
          in: path
          required: true
          description: Unique identifier of the job
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job results
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                  results:
                    type: object
                  processing_time_ms:
                    type: integer
                  completed_at:
                    type: string
                    format: date-time
        '202':
          description: Job still processing
        '404':
          description: Job not found or no results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/jobs/{job_id}:
    delete:
      summary: Cancel job
      description: Cancel a running or queued job
      tags:
        - jobs
      parameters:
        - name: job_id
          in: path
          required: true
          description: Unique identifier of the job to cancel
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  job_id:
                    type: string
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Job cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    SimilaritySearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The search query text
          example: "How to configure the database connection?"
        collection_id:
          type: string
          description: Collection ID to search in
          default: "default"
        limit:
          type: integer
          description: Maximum number of results to return
          default: 10
          minimum: 1
          maximum: 100
        min_score:
          type: number
          format: float
          description: Minimum similarity score threshold
          minimum: 0.0
          maximum: 1.0
        metadata_filter:
          type: object
          additionalProperties: true
          description: Metadata filters to apply
    SearchResult:
      type: object
      required:
        - document_id
        - content
        - score
        - metadata
      properties:
        document_id:
          type: string
          format: uuid
          description: Unique identifier of the document
        content:
          type: string
          description: The matching content/chunk
        score:
          type: number
          format: float
          description: Similarity score (0-1)
        metadata:
          type: object
          additionalProperties: true
          description: Document metadata
        chunk_index:
          type: integer
          description: Index of the chunk within the document
    SimilaritySearchResponse:
      type: object
      required:
        - query
        - results
        - total_found
        - processing_time_ms
      properties:
        query:
          type: string
          description: The original search query
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
          description: List of search results
        total_found:
          type: integer
          description: Total number of results found
        processing_time_ms:
          type: integer
          description: Time taken to process the search in milliseconds
...
    BatchSearchRequest:
      type: object
      required:
        - queries
      properties:
        queries:
          type: array
          items:
            type: string
          description: List of search queries
        collection_id:
          type: string
          description: Collection ID to search in
          default: "default"
        limit:
          type: integer
          description: Maximum number of results per query
          default: 10
          minimum: 1
          maximum: 100
        output_format:
          type: string
          enum: [json, csv]
          description: Output format
          default: "json"
        metadata_filter:
          type: object
          additionalProperties: true
          description: Metadata filters to apply
    BatchSearchResponse:
      type: object
      required:
        - job_id
        - queries_processed
        - results
        - processing_time_ms
        - status
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique identifier for the batch job
        queries_processed:
          type: integer
          description: Number of queries processed
        results:
          type: array
          items:
            $ref: '#/components/schemas/SimilaritySearchResponse'
          description: List of search results for each query
        processing_time_ms:
          type: integer
          description: Total time taken to process all queries
        status:
          type: string
          enum: [completed, failed, processing]
          description: Status of the batch search
...
    CollectionInfo:
      type: object
      required:
        - id
        - name
        - document_count
        - embedding_count
        - created_at
        - updated_at
        - metadata
      properties:
        id:
          type: string
          description: Unique identifier of the collection
        name:
          type: string
          description: Name of the collection
        description:
          type: string
          description: Description of the collection
        document_count:
          type: integer
          description: Number of documents in the collection
        embedding_count:
          type: integer
          description: Number of embeddings in the collection
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        metadata:
          type: object
          additionalProperties: true
          description: Collection metadata
...
    CollectionListResponse:
      type: object
      required:
        - collections
        - total_count
      properties:
        collections:
          type: array
          items:
            $ref: '#/components/schemas/CollectionInfo'
          description: List of collections
        total_count:
          type: integer
          description: Total number of collections
...
    JobStatus:
      type: object
      required:
        - job_id
        - status
        - progress
        - started_at
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique identifier of the job
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
          description: Current status of the job
        progress:
          type: number
          format: float
          description: Job progress (0-1)
          minimum: 0.0
          maximum: 1.0
        started_at:
          type: string
          format: date-time
          description: Job start timestamp
        completed_at:
          type: string
          format: date-time
          description: Job completion timestamp
        error_message:
          type: string
          description: Error message if job failed
        result_url:
          type: string
          description: URL to retrieve job results
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        request_id:
          type: string
          description: Request ID for debugging
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - BearerAuth: []
...