# src/api/jobs.py
from fastapi import APIRouter, HTTPException, status
from typing import Dict, Any
import logging
from datetime import datetime

from .models import (
    JobStatus,
    ErrorResponse
)

logger = logging.getLogger(__name__)
router = APIRouter()


@router.get("/jobs/{job_id}/status",
            response_model=JobStatus,
            responses={
                404: {"model": ErrorResponse, "description": "Job not found"},
                500: {"model": ErrorResponse, "description": "Internal server error"}
            })
async def get_job_status(job_id: str) -> JobStatus:
    """
    Get the status of an asynchronous job.
    
    Returns information about job progress, completion status,
    and any error messages if the job failed.
    """
    try:
        logger.info(f"Retrieving job status for: {job_id}")
        
        # TODO: Implement actual job status tracking
        # For now, return mock status
        
        # Mock job status
        mock_status = JobStatus(
            job_id=job_id,
            status="completed",  # Could be: queued, running, completed, failed
            progress=1.0,
            started_at=datetime.now(),
            completed_at=datetime.now(),
            error_message=None,
            result_url=f"/api/v1/jobs/{job_id}/results"
        )
        
        return mock_status
        
    except Exception as e:
        logger.error(f"Job status retrieval failed: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Job status retrieval failed: {str(e)}"
        )


@router.get("/jobs/{job_id}/results",
            responses={
                200: {"description": "Job results"},
                404: {"model": ErrorResponse, "description": "Job not found or no results"},
                202: {"description": "Job still processing"},
                500: {"model": ErrorResponse, "description": "Internal server error"}
            })
async def get_job_results(job_id: str) -> Dict[str, Any]:
    """
    Retrieve the results of a completed job.
    
    Returns the actual results of the job if completed,
    or an error if the job is still processing or failed.
    """
    try:
        logger.info(f"Retrieving job results for: {job_id}")
        
        # TODO: Implement actual job results retrieval
        # For now, return mock results
        
        # Mock job results
        mock_results = {
            "job_id": job_id,
            "status": "completed",
            "results": {
                "documents_processed": 10,
                "embeddings_created": 50,
                "success_count": 10,
                "error_count": 0
            },
            "processing_time_ms": 5000,
            "completed_at": datetime.now().isoformat()
        }
        
        return mock_results
        
    except Exception as e:
        logger.error(f"Job results retrieval failed: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Job results retrieval failed: {str(e)}"
        )


@router.delete("/jobs/{job_id}",
               responses={
                   200: {"description": "Job cancelled successfully"},
                   404: {"model": ErrorResponse, "description": "Job not found"},
                   409: {"model": ErrorResponse, "description": "Job cannot be cancelled"},
                   500: {"model": ErrorResponse, "description": "Internal server error"}
               })
async def cancel_job(job_id: str) -> Dict[str, Any]:
    """
    Cancel a running or queued job.
    
    Attempts to cancel the specified job if it's still in progress.
    Completed jobs cannot be cancelled.
    """
    try:
        logger.info(f"Cancelling job: {job_id}")
        
        # TODO: Implement actual job cancellation logic
        # This would:
        # 1. Check if job exists
        # 2. Check if job can be cancelled
        # 3. Stop job processing
        # 4. Clean up partial results
        # 5. Update job status
        
        return {
            "message": f"Job {job_id} cancelled successfully",
            "job_id": job_id,
            "status": "cancelled",
            "timestamp": datetime.now().isoformat()
        }
        
    except Exception as e:
        logger.error(f"Job cancellation failed: {str(e)}")
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Job cancellation failed: {str(e)}"
        )